+++ frontend/src/routes/AppRoutes.jsx
@@ -1,10 +1,11 @@
 import { Routes, Route } from 'react-router-dom';
 
 import ProtectedRoute from '@components/common/ProtectedRoute/ProtectedRoute';
-import HomePage from '@pages/HomePage/HomePage';
-import LoginPage from '@pages/auth/Login/LoginPage';
+
+import HomePage from '@pages/home/HomePage';
+import LoginPage from '@pages/login/LoginPage';
 import DashboardPage from '@pages/dashboard/DashboardPage';
-import NotFoundPage from '@pages/NotFoundPage/NotFoundPage';
+import NotFoundPage from '@pages/not-found/NotFoundPage';
 
 // TODO: Import create and import ProtectedRoute component
